requirements.txt
torch
transformers
gradio
python-dotenv

